fix(webpack.plugins.ts): use SERVER_URL on the client side to connect to API
The meticulous backend which is offering the server API for the dial app is usually running on
the same machine as the dail. However in certain cases we want to be able to change this, either for
development purposes or when we want the dial to be able to be displayed on the users phone for
accessibility reasons later.
The documentation calls for setting SERVER_URL however it is never actually used. Export it through
the well known process.env interface which is usually not available in setups like ours where we are
running a frontend application. To still use this we are changing the config to export the env as an
additional plugin.

Signed-off-by: Mimoja <mimoja@meticuloushome.com>