fix(circle-keyboard): don't encode special characters as html entities
The circlular keyboards special characters are currently encoded as strings
containing html entities. This leads to the profile name sometimes rendering
incorrectly but even worse: Transfering to the profile json broken.
To fix this we are instead using the 1992s technology UTF-8 which is universally
supported across our stack.
While we are converting the representation we are also replacing the multiplication
sign (×) and the division sign (÷) with the german Eszett or sharp S (ß and ẞ).

Signed-off-by: Mimoja <mimoja@meticuloushome.com>