fix(presets): always clear the animating option even indenpendant of profile source
Issue:
When selecting and starting a profile from the mobile app the dial app would never
leave the animating state after returning to the Pressets.tsx screen and therefore
refuse all dial inputs.
We force a clearing of the animating state with an useEffect which helps with timeouts
overlapping multiple redraws.

Signed-off-by: Mimoja <git@mimoja.de>