fix(profile settings): added support for new variables
The variables of type "piston position" and "motor power" did not have a gauge
screen available for edition, so they could not be edited in the dial.This commit
adds the screens necessary to support them.

As a "motor power" variable can be set to negative values, support for negative values
on the gauge is added as well as a "minimum value" to be set for each setting screen to
replace the 0 but keeping it as the default minimum.