feat(settings/advanced): custrom piston retraction
A new option 'Shot volume' is added to the advanced settings menu, with
it You can set the retraction distance the piston follows to let it
the selected volume of water to the cylinder.

The volume range goes from 100ml to 150ml. That value is purely visual and
used to calculate the retraction distance (mm) for the piston to be saved
in the settings menu.