chore(purgescreen): show last shot graph on purge
The latest shot graph is shown while the purge at the end of it
is happening, the piston animation size got reduced.

If the graph is not interacted with while purging, at its end, it
returns to the profile home screen, if interacted with, the graph
is kept until a click is done.

The inners of ShotGraphScreen got refaactored into a ShotGraph component
to take out unnecesary gestures and be able to embed it into
different screens, it now takes as props the `profile` name to get
the latest data from, and the `isStatic` flag to disable interactions
with it