fix(profilehomescreen): images not loading
There is an issue of syncronization between the "scroll" animation and
the loading of the profile images, this causes that the image
containers get unmounted while the animation is still on going and in
the space of the now unmounted image, thats why it only renders the
profile container with the colo(u)r backgound.

Now when mounting the image container we mark it as unsafe to remove,
observe if it intersects the viewport (is visible) and when it is no
longer  visible we remove the mark allowing it to be unmounted.