fix(quicksettings): exit option was transparent to click
When clicking the 'exit' option in the quick settings menu, the gesture
was also registered on the underlying screen.

This was probably because when handling the 'gesture' event, it was handled
first by the quick settings menu, updating the 'visible' attribute of the
bubbleDisplay to false before the underlying screen handle its gesture, so by
the time it does, the gesture is also handled there.

We now implement a 'interceptsGesture' attribute to the screen state
under bubbledisplay. This way the 'visible' attribute only handles the
visibility of the component, while 'interceptGesture' is used in
'useHandleGestures' to not consume the gesture in the underlying screens

With that new implementation, when clicking on the 'exit' option first the
visibility is toggled off, and then after 100ms timeout, the 'intercerptGesture'
is turned off.