fix(profilehomescreen): lock profile displacement while pressing down knob
Issue:
When the knob is pressed down to start the profile brew we can still move
through the profile list. This can cause either to brew a non intended profile,
or if ends up in the "new profile" option, crash the dial app.

This commit locks the movement of the profile list while the knob is pressed down